Begins to implement the blog posting
Added a 'create blog post' option in the main navbar. Checks to ensure the logged in
user is a blogPoster before showing that link. Only has a text field implemented for
now. The next step is to set up some views to actually see the blogs posts then work
on the markdown conversion.