Implements html file upload capability
Previously I had to load the an html file that I converted myself with pandoc
manually to the website files. Now the user can upload a file though the site.
I guess the next step would be to let the user upload a markdown file and
convert it on the site into html.