Adds a database backup and reconfigures .gitignore for it
I have handled the database in my git very poorly. I had been thinking that
I could just delete it every time and when I clone
or pull from github jsut recreate the db. That is annoying
because then I have to manually re-enter example data. Even more
problematic is that no longer creates tables for my apps. So
I need to manually call makemigrations 'app'. I was also
stupidly deleting migration folders in the apps which was
probably the cause of this issue.

Now I dump the database into 'whole_db.sql' which is added to the
repo. I also included the app migration folders in the repo. Now
all I have to do when I clone or fetch/pull is restore the
local database with the dump file.