Removes the deprecated blogPost app
The blogPost app has been replaced with the content app. I designed
blogPost to allow any users to create and manage their posts. As
the website progressed I realized that was unecessary because I want
to be the only one posting. It seemed easier to just start a new
simpler app that I named content to replace blogPost.