Introduces 'include' tag to the detail view
I have been struggling for a while now to incorporate these tutorial
html files, which I ~ create by hand through markdown and pandoc,
into the detail view here. The solution is to use the include template
tag within the post_detail.html.

For now I just typed an example tutorial html file in by hand. Going
forward I can add a field to the ContentPost model to store
the path to the associated template file (i.e. the real tutorial
html that I made) then use that in the include tag.

I am still on the fence about if I want to add these tutorial html
files to the model as a fileField. That would necessitate the
include tag looking outside of the template root folder. Probably
doable but how automated does it really need to be?