Reverts to handling articles without git gists
I decided that I dont like how those gists look, if I handle
the post template files myself then I can customize them to
look however I want.

I am doing things differently this time. I created a FileField
on the ContentPost model to hold the article template files (really just an
html file mostly but I want to have the ability to include template
tags and variables if need be). This way everything is stored in media, I
can edit it if need be and I don't have to worry about maintaining them
by hand as much.

Currently the html text file of each article is converted to a Template
and then included in the detail view for its object. This works fine
but seems a little wastful. In the next step I will make a TemplateField,
if there is such a thing, in the PostContent model. Then I can create the
template in the save() function of the model and just do it once, rather
than doing it everytime the view is hit.