Permits post creation dates to be edited
Originally I thought it would be fine to just leave it as whenever the
ContentPost instance is first saved. This doesn't work well in deployment,
at least at this point when I am frequently deleting things
from the database.

I changed to to default to the current date but it can also be changed
now.