Orders the article lists by creation date
This commit also limits the number of posts on the landing page
to 4, since the list title is 'most recent'.

Further in this commit I decided to only implement the oredering in the
LandingPageView and then make the ContentPostListView a
subclass of it so I only needed to write the ordering context
stuff once.