Major overhaul of the article lists
I found with the previous setup that when I tried to change
the image size everything went to hell. The text became
cut-off or unreadable. So I have redone the article lists
with bootstrap now. It works well on phones too.

There are still still some minor styling things to work on
probably but I will revisit them later.